A version of Sinkhorn OT on distributions.
This solves the problem that we previously had when transporting
non-overlapping distributions -as-functions: the Haar representation
can't describe the zero regions very accurately, because they are
constructed as an integral offset minus local corrections.

The problem does not arise with distributions-as-dual vectors, because
here localised peaks don't correspond to high amplitudes.